moving logical things from protocol to utils redoing imports
