+++ helixdb/src/protocol/mod.rs
@@ -1,9 +1,4 @@
-pub mod count;
 pub mod date;
-pub mod filterable;
-pub mod id;
-pub mod items;
-pub mod label_hash;
 pub mod remapping;
 pub mod request;
 pub mod response;
