+++ helixdb/src/protocol/return_values.rs
@@ -1,9 +1,11 @@
 use super::{
+    remapping::{Remapping, ResponseRemapping},
+    value::Value,
+};
+use crate::utils::{
     count::Count,
     filterable::{Filterable, FilterableType},
     items::{Edge, Node},
-    remapping::{Remapping, ResponseRemapping},
-    value::Value,
 };
 use crate::helix_engine::graph_core::ops::tr_val::TraversalVal;
 use sonic_rs::{Deserialize, Serialize};
