+++ helixdb/src/utils/count.rs
