+++ helixdb/src/utils/id.rs
