+++ helixdb/src/utils/mod.rs
@@ -1,2 +1,6 @@
 pub mod styled_string;
-
+pub mod count;
+pub mod filterable;
+pub mod id;
+pub mod items;
+pub mod label_hash;
